+++ .env
@@ -1,3 +1,3 @@
 DB_PASS=Lhx970201
-spoonAPI=487f50a6f3msh5e994b878b01689p13c870jsn00952d320107
+SPOONAPI=487f50a6f3msh5e994b878b01689p13c870jsn00952d320107
 PASSW_DW=4eb5968d74msh7a28f9eece9c17ap13c40djsnde5aaab47e32
