+++ public/htmlExamples/menu.html
@@ -21,7 +21,7 @@
       }
         .container{
         float: center;
-        width: 70%;
+        /* width: 70%; */
       }
      </style>
      <title>
