+++ utils/api_music.js
@@ -1,6 +1,6 @@
 /* Load the HTTP library */
 var unirest = require("unirest");
-require("dotenv").config()
+require("dotenv").config({path:"../.env"})
 
 
 var req = unirest("GET", "https://deezerdevs-deezer.p.rapidapi.com/search");
